feat(continuity): implement request continuity
This patch adds a module which is capable of aborting a request, waiting for
another option to complete, and then resuming the request again.

This feature makes use of a temporary cookie which keeps track of the
request state.

This feature is required for several workflows that update privileged
fields such as passwords, 2fa recovery codes, email addresses.